requirements.txt
requests>=2.28
google-cloud-storage>=2.8
python-dateutil>=2.8
urllib3>=1.26
